# Refactor `transition_range` and callers