--- a/src/TimeZones.jl
+++ b/src/TimeZones.jl
@@ -33,6 +33,10 @@ export TimeZone, @tz_str, istimezone, FixedTimeZone, VariableTimeZone, ZonedDate
 const PKG_DIR = normpath(joinpath(dirname(@__FILE__), ".."))
 const DEPS_DIR = joinpath(PKG_DIR, "deps")
 
+# TimeZone types used to disambiguate the context of a DateTime
+# abstract type UTC <: TimeZone end  # Already defined in the Dates stdlib
+abstract type Local <: TimeZone end
+
 function __init__()
     # Base extension needs to happen everytime the module is loaded (issue #24)
     Dates.CONVERSION_SPECIFIERS['z'] = TimeZone