--- a/test/interpret.jl
+++ b/test/interpret.jl
@@ -1,3 +1,84 @@
+using Dates: Millisecond
+using TimeZones: transition_range
+
+@testset "lt_local / lt_utc" begin
+    t = Transition(DateTime(1977, 5, 26, 1), FixedTimeZone("PDT", -8 * 3600, 3600))
+    ms = Millisecond(1)
+
+    @testset "lt_local" begin
+        local_dt = DateTime(1977, 5, 25, 18)
+
+        @test TimeZones.lt_local(local_dt - ms, t)
+        @test !TimeZones.lt_local(local_dt, t)
+        @test !TimeZones.lt_local(local_dt + ms, t)
+
+        @test !TimeZones.lt_local(t, local_dt - ms)
+        @test !TimeZones.lt_local(t, local_dt)
+        @test TimeZones.lt_local(t, local_dt + ms)
+    end
+
+    @testset "lt_utc" begin
+        utc_dt = DateTime(1977, 5, 26, 1)
+
+        @test TimeZones.lt_utc(utc_dt - ms, t)
+        @test !TimeZones.lt_utc(utc_dt, t)
+        @test !TimeZones.lt_utc(utc_dt + ms, t)
+
+        @test !TimeZones.lt_utc(t, utc_dt - ms)
+        @test !TimeZones.lt_utc(t, utc_dt)
+        @test TimeZones.lt_utc(t, utc_dt + ms)
+    end
+end
+
+@testset "transition_range" begin
+    # TODO: Redundancy with test from ZonedDateTime
+    @testset "multiple hour transitions" begin
+        # Transitions changes that exceed an hour. Results in having two sequential
+        # non-existent hour and two sequential ambiguous hours.
+        tz = VariableTimeZone("Testing", [
+            Transition(DateTime(1800,1,1), FixedTimeZone("TST",0,0)),
+            Transition(DateTime(1950,4,1), FixedTimeZone("TDT",0,7200)),
+            Transition(DateTime(1950,9,1), FixedTimeZone("TST",0,0)),
+        ])
+
+        ### Local ###
+
+        # Initial transition
+        @test transition_range(DateTime(1799, 12, 31, 23), tz, Local) == 1:0
+        @test transition_range(DateTime(1800, 01, 01, 00), tz, Local) == 1:1
+
+        # A "spring forward" where 2 hours are skipped.
+        @test transition_range(DateTime(1950, 03, 31, 23), tz, Local) == 1:1
+        @test transition_range(DateTime(1950, 04, 01, 00), tz, Local) == 2:1  # 00:00 TST/TDT
+        @test transition_range(DateTime(1950, 04, 01, 01), tz, Local) == 2:1  # 01:00 TST/TDT
+        @test transition_range(DateTime(1950, 04, 01, 02), tz, Local) == 2:2
+
+        # A "fall back" where 2 hours are duplicated. Never appears to occur in reality.
+        @test transition_range(DateTime(1950, 08, 31, 23), tz, Local) == 2:2
+        @test transition_range(DateTime(1950, 09, 01, 00), tz, Local) == 2:3  # 00:00 TDT/TST
+        @test transition_range(DateTime(1950, 09, 01, 01), tz, Local) == 2:3  # 01:00 TDT/TST
+        @test transition_range(DateTime(1950, 09, 01, 02), tz, Local) == 3:3
+
+        ### UTC ###
+
+        # Initial transition
+        @test transition_range(DateTime(1799, 12, 31, 23), tz, UTC) == 1:0
+        @test transition_range(DateTime(1800, 01, 01, 00), tz, UTC) == 1:1
+
+        # A "spring forward" where 2 hours are skipped.
+        @test transition_range(DateTime(1950, 03, 31, 23), tz, UTC) == 1:1  # 23:00 TST
+        @test transition_range(DateTime(1950, 04, 01, 00), tz, UTC) == 2:2  # 02:00 TDT
+
+        # A "fall back" where 2 hours are duplicated. Never appears to occur in reality.
+        @test transition_range(DateTime(1950, 08, 31, 21), tz, UTC) == 2:2
+        @test transition_range(DateTime(1950, 08, 31, 22), tz, UTC) == 2:2  # 00:00 TDT
+        @test transition_range(DateTime(1950, 08, 31, 23), tz, UTC) == 2:2  # 01:00 TDT
+        @test transition_range(DateTime(1950, 09, 01, 00), tz, UTC) == 3:3  # 00:00 TST
+        @test transition_range(DateTime(1950, 09, 02, 00), tz, UTC) == 3:3  # 01:00 TST
+        @test transition_range(DateTime(1950, 09, 03, 00), tz, UTC) == 3:3
+    end
+end
+
 # Contains both positive and negative UTC offsets and observes daylight saving time.
 apia = first(compile("Pacific/Apia", tzdata["australasia"]))