--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -42,6 +42,8 @@ include("helpers.jl")
 
 @testset "TimeZones" begin
     include("utils.jl")
+    include("indexable_generator.jl")
+
     include("class.jl")
     include(joinpath("tzdata", "timeoffset.jl"))
     include(joinpath("tzdata", "archive.jl"))