@omus 2020-09-17
Increasing code coverage. Also removed no longer needed imports.