@omus 2020-09-18
I'm performing some final validation (I definitely need to work on formalizing benchmarks for this package next) and have discovered a regression.

On Julia 1.5.1:

Current PR:
```julia
julia> @benchmark collect($(ZonedDateTime(2020, tz"America/Winnipeg"):Hour(1):ZonedDateTime(2021, tz"America/Winnipeg")))
BenchmarkTools.Trial:
  memory estimate:  1.27 MiB
  allocs estimate:  26381
  --------------
  minimum time:     1.054 ms (0.00% GC)
  median time:      1.361 ms (0.00% GC)
  mean time:        1.574 ms (5.58% GC)
  maximum time:     9.904 ms (0.00% GC)
  --------------
  samples:          3175
  evals/sample:     1
```
Current master (48fe988):
```julia
julia> @benchmark collect($(ZonedDateTime(2020, tz"America/Winnipeg"):Hour(1):ZonedDateTime(2021, tz"America/Winnipeg")))
BenchmarkTools.Trial:
  memory estimate:  1.81 MiB
  allocs estimate:  17588
  --------------
  minimum time:     787.054 μs (0.00% GC)
  median time:      1.169 ms (0.00% GC)
  mean time:        1.978 ms (6.01% GC)
  maximum time:     107.202 ms (0.00% GC)
  --------------
  samples:          2526
  evals/sample:     1
```
The increase in allocations is especially interesting. I'll try to get to the bottom of this before proceeding.