@omus 2020-09-18
I've manged to address the regression and improve things further. It appears that the introduction of a function barrier when using a `FixedTimeZone` resulted in the slowdown and increased allocations:

Current PR (1ed4d09):
```julia
julia> @benchmark collect($(ZonedDateTime(2020, tz"America/Winnipeg"):Hour(1):ZonedDateTime(2021, tz"America/Winnipeg")))
BenchmarkTools.Trial:
  memory estimate:  755.44 KiB
  allocs estimate:  8795
  --------------
  minimum time:     444.120 μs (0.00% GC)
  median time:      683.956 μs (0.00% GC)
  mean time:        859.080 μs (6.40% GC)
  maximum time:     15.121 ms (0.00% GC)
  --------------
  samples:          5815
  evals/sample:     1
```

Current master (48fe988):
```julia
julia> @benchmark collect($(ZonedDateTime(2020, tz"America/Winnipeg"):Hour(1):ZonedDateTime(2021, tz"America/Winnipeg")))
BenchmarkTools.Trial:
  memory estimate:  1.81 MiB
  allocs estimate:  17588
  --------------
  minimum time:     788.836 μs (0.00% GC)
  median time:      998.876 μs (0.00% GC)
  mean time:        1.210 ms (5.79% GC)
  maximum time:     25.093 ms (0.00% GC)
  --------------
  samples:          4128
  evals/sample:     1
```
Initially posted benchmarks in the description are still accurate.